added code to list all Os & App deps
