+++ pkg/types/vulnerability.go
@@ -1,17 +1,14 @@
 package types
 
-import (
-	ftypes "github.com/aquasecurity/fanal/types"
-	"github.com/aquasecurity/trivy-db/pkg/types"
-)
-
-type DetectedVulnerability struct {
-	VulnerabilityID  string       `json:",omitempty"`
+type OsPackage struct {
 	PkgName          string       `json:",omitempty"`
 	InstalledVersion string       `json:",omitempty"`
-	FixedVersion     string       `json:",omitempty"`
-	Layer            ftypes.Layer `json:",omitempty"`
-	SeveritySource   string       `json:",omitempty"`
 
-	types.Vulnerability
+	Package
+}
+
+type Package struct {
+	Title          string         `json:",omitempty"`
+	Description    string         `json:",omitempty"`
+	References     []string       `json:",omitempty"`
 }
